[EPMEDU-1269]: Clean Architecture Adjustments
