+++ feature/home/src/main/java/com/epmedu/animeal/home/di/HomeDataModule.kt
@@ -1,8 +1,8 @@
 package com.epmedu.animeal.home.di
 
 import com.epmedu.animeal.common.component.AppSettingsProvider
-import com.epmedu.animeal.home.data.ApplicationSettingsRepository
 import com.epmedu.animeal.home.data.ApplicationSettingsRepositoryImpl
+import com.epmedu.animeal.home.domain.ApplicationSettingsRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
