+++ feature/home/src/main/java/com/epmedu/animeal/home/di/HomeDomainModule.kt
@@ -1,7 +1,7 @@
 package com.epmedu.animeal.home.di
 
 import com.epmedu.animeal.feeding.domain.repository.FeedingPointRepository
-import com.epmedu.animeal.home.data.ApplicationSettingsRepository
+import com.epmedu.animeal.home.domain.ApplicationSettingsRepository
 import com.epmedu.animeal.home.domain.usecases.CancelFeedingUseCase
 import com.epmedu.animeal.home.domain.usecases.FinishFeedingUseCase
 import com.epmedu.animeal.home.domain.usecases.GetAllFeedingPointsUseCase
