+++ feature/home/src/main/java/com/epmedu/animeal/home/domain/usecases/GetGeolocationPermissionRequestedSettingUseCase.kt
@@ -1,6 +1,6 @@
 package com.epmedu.animeal.home.domain.usecases
 
-import com.epmedu.animeal.home.data.ApplicationSettingsRepository
+import com.epmedu.animeal.home.domain.ApplicationSettingsRepository
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
