+++ feature/signupflow/entercode/src/main/java/com/epmedu/animeal/signup/entercode/di/EnterCodeModule.kt
@@ -3,8 +3,8 @@ package com.epmedu.animeal.signup.entercode.di
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import com.epmedu.animeal.auth.AuthAPI
-import com.epmedu.animeal.signup.entercode.data.EnterCodeRepository
 import com.epmedu.animeal.signup.entercode.data.EnterCodeRepositoryImpl
+import com.epmedu.animeal.signup.entercode.domain.EnterCodeRepository
 import com.epmedu.animeal.signup.entercode.domain.FacebookConfirmCodeUseCase
 import com.epmedu.animeal.signup.entercode.domain.GetPhoneNumberUseCase
 import com.epmedu.animeal.signup.entercode.domain.MobileConfirmCodeUseCase
