+++ feature/signupflow/entercode/src/main/java/com/epmedu/animeal/signup/entercode/domain/FacebookConfirmCodeUseCase.kt
@@ -1,7 +1,6 @@
 package com.epmedu.animeal.signup.entercode.domain
 
 import com.epmedu.animeal.auth.AuthRequestHandler
-import com.epmedu.animeal.signup.entercode.data.EnterCodeRepository
 
 class FacebookConfirmCodeUseCase(private val repository: EnterCodeRepository) {
