+++ feature/signupflow/entercode/src/main/java/com/epmedu/animeal/signup/entercode/domain/GetPhoneNumberUseCase.kt
@@ -1,6 +1,5 @@
 package com.epmedu.animeal.signup.entercode.domain
 
-import com.epmedu.animeal.signup.entercode.data.EnterCodeRepository
 import kotlinx.coroutines.flow.Flow
 
 class GetPhoneNumberUseCase(private val repository: EnterCodeRepository) {
