+++ feature/signupflow/entercode/src/main/java/com/epmedu/animeal/signup/entercode/domain/SendCodeUseCase.kt
@@ -1,7 +1,6 @@
 package com.epmedu.animeal.signup.entercode.domain
 
 import com.epmedu.animeal.auth.AuthRequestHandler
-import com.epmedu.animeal.signup.entercode.data.EnterCodeRepository
 import java.lang.Exception
 
 class SendCodeUseCase(private val repository: EnterCodeRepository) {
