+++ feature/signupflow/enterphone/src/main/java/com/epmedu/animeal/signup/enterphone/domain/SignUpAndSignInUseCase.kt
@@ -2,7 +2,6 @@ package com.epmedu.animeal.signup.enterphone.domain
 
 import com.amplifyframework.auth.cognito.exceptions.service.UsernameExistsException
 import com.epmedu.animeal.auth.AuthRequestHandler
-import com.epmedu.animeal.signup.enterphone.data.EnterPhoneRepository
 
 class SignUpAndSignInUseCase(private val repository: EnterPhoneRepository) {
     operator fun invoke(
