+++ shared/feature/profile/src/main/java/com/epmedu/animeal/profile/domain/GetProfileUseCase.kt
@@ -1,6 +1,6 @@
 package com.epmedu.animeal.profile.domain
 
-import com.epmedu.animeal.profile.data.repository.ProfileRepository
+import com.epmedu.animeal.profile.domain.repository.ProfileRepository
 
 class GetProfileUseCase(private val repository: ProfileRepository) {
     suspend operator fun invoke() = repository.getProfile()
